Optimized RGB Arrays to be uint8

+++ screenController/colorInterpolator.h
@@ -3,30 +3,30 @@
 #include "colorProvider.h"
 
 typedef struct {
-    double r;       // a fraction between 0 and 1
-    double g;       // a fraction between 0 and 1
-    double b;       // a fraction between 0 and 1
+    uint8_t r;       // a fraction between 0 and 1
+    uint8_t g;       // a fraction between 0 and 1
+    uint8_t b;       // a fraction between 0 and 1
 } rgb;
 
 typedef struct {
-    double h;       // angle in degrees
-    double s;       // a fraction between 0 and 1
-    double v;       // a fraction between 0 and 1
+    uint8_t h;       // angle in degrees
+    uint8_t s;       // a fraction between 0 and 1
+    uint8_t v;       // a fraction between 0 and 1
 } hsv;
 
 class ColorInterpolator : public ColorProvider{
     public:
       // Takes in color1, color2 in RGB format
-      ColorInterpolator(double* color1, double* color2);
+      ColorInterpolator(uint8_t* color1, uint8_t* color2);
       // getColor() returns the color value in RGB format
-      virtual double* getColor();
+      virtual uint8_t* getColor();
     private:
       hsv startColor;
       hsv endColor;
-      double t;
+      uint8_t t;
       hsv rgb2hsv(rgb in);
       rgb hsv2rgb(hsv in);
-      double interpolateLinearly(double a, double b, double t);
+      uint8_t interpolateLinearly(uint8_t a, uint8_t b, uint8_t t);
 };
 
 #endif
